Rest Callout working
Rest call is working. just need to get the exact information im looking for (name / image URL) and make the callout several times.  once for each pokemon_x__c returned.

after that, go back to the VF page and display the images of each pokemon and the name to go with them.